0.2.15

+++ code/actorComponents/body.cpp
@@ -1,4 +1,4 @@
-c_body::c_body(c_actor* father, const bool& canMove, const bool& canView, const bool& canPickup, const float& weight)
+c_body::c_body(c_actor* father, const bool& canMove, const bool& canView, const bool& canPickup, const int& weight)
 : father(father),
   canMove(canMove),
   canView(canView),
